test-requirements.txt
# Hacking already pins down pep8, pyflakes and flake8
hacking>=0.9.2,<0.10

# fix version problems
cffi>0.9
oslo.i18n==1.7.0
oslo.utils==1.9.0
oslo.serialization==1.7.0

coverage>=3.6
django-nose
mock>=1.0
funcsigs
mox3>=0.7.0
nodeenv
nose
nose-exclude
nosexcover
openstack.nose_plugin>=0.7
nosehtmloutput>=0.0.3
selenium
# Docs Requirements
sphinx>=1.1.2,!=1.2.0,!=1.3b1,<1.3
oslosphinx
http://tarballs.openstack.org/horizon/horizon-master.tar.gz#egg=horizon
